NEW - bug 193789: Failed migrate old data folder https://bugs.eclipse.org/bugs/show_bug.cgi?id=193789

+++ org.eclipse.mylyn.tasks.ui/src/org/eclipse/mylyn/internal/tasks/ui/WorkspaceAwareContextStore.java
@@ -23,8 +23,6 @@ import org.eclipse.mylyn.tasks.ui.TasksUiPlugin;
  */
 public class WorkspaceAwareContextStore extends AbstractContextStore {
 
-	private static final String MIGRATION_FAILED = "Failed migrate old data folder";
-
 	private static final String DIRECTORY_METADATA = ".metadata";
 
 	private static final String OLD_DATA_DIR = ".mylar";
@@ -37,22 +35,30 @@ public class WorkspaceAwareContextStore extends AbstractContextStore {
 
 	@Override
 	public void init() {
-		// Migrate .mylar data folder to .metadata/mylyn
-		// if user was still using default location
+		// Migrate .mylar data folder to .metadata/.mylyn
 		String oldDefaultDataPath = ResourcesPlugin.getWorkspace().getRoot().getLocation().toString() + '/'
 				+ OLD_DATA_DIR;
+		File newDefaultDataDir = new File(TasksUiPlugin.getDefault().getDefaultDataDirectory());
 		File oldDefaultDataDir = new File(oldDefaultDataPath);
-		if (oldDefaultDataDir.exists()) {
-				File metadata = new File(ResourcesPlugin.getWorkspace().getRoot().getLocation().toString() + '/'
-						+ DIRECTORY_METADATA);
-				if (!metadata.exists()) {
-					StatusManager.log(MIGRATION_FAILED, this);
-				} else {
-					if (!oldDefaultDataDir.renameTo(new File(TasksUiPlugin.getDefault().getDefaultDataDirectory()))) {
-						StatusManager.log(MIGRATION_FAILED, this);
-					}
+		if (newDefaultDataDir.exists() && oldDefaultDataDir.exists()) {
+			StatusManager.log("Legacy data folder detected: " + oldDefaultDataDir.getAbsolutePath(), this);
+		} else if (oldDefaultDataDir.exists() && !newDefaultDataDir.exists()) {
+			File metadata = new File(ResourcesPlugin.getWorkspace().getRoot().getLocation().toString() + '/'
+					+ DIRECTORY_METADATA);
+			if (!metadata.exists()) {
+				if (!metadata.mkdirs()) {
+					StatusManager.log("Unable to create metadata folder: " + metadata.getAbsolutePath(), this);
 				}
+			}
+
+			if (metadata.exists()) {
+				if (!oldDefaultDataDir.renameTo(new File(TasksUiPlugin.getDefault().getDefaultDataDirectory()))) {
+					StatusManager.log("Failed to migrate legacy data from " + oldDefaultDataDir.getAbsolutePath()
+							+ " to " + TasksUiPlugin.getDefault().getDefaultDataDirectory(), this);
+				}
+			}
 		}
+
 		rootDirectory = new File(TasksUiPlugin.getDefault().getDataDirectory());
 		if (!rootDirectory.exists()) {
 			rootDirectory.mkdir();
